r=int(input()) Area=3.14*r**2 print(Area)

+++ Area_of_a_circle.py
@@ -1,3 +1,3 @@
-r=float(input())
+r=int(input())
 Area=3.14*r**2
-print( Area)
+print(Area)
